.github/workflows/requirements.txt
google-generativeai
requests
python-dotenv
pymongo[srv]==3.11
